+++ client/src/components/drawer/MiniDrawer.js
@@ -186,7 +186,7 @@ export default function MiniDrawer({ element }) {
               {
                 id: 14,
                 name: "Coding Agile",
-                route: "#",
+                route: "/agile",
                 icon: (
                   <TimelineIcon
                     sx={{ color: theme.palette.custom.iconDrawer }}
