+++ server/server.js
@@ -77,7 +77,8 @@ ws.on("request", (request) => {
   require("./routes/call")(app, db, connection, pathname);
   require("./dashboard")(app, db, ws, parse);
   require("./routes/groupscreation")(app, db);
-  require("./userInfo")(app, pathname, db, connection, upload,path,fs);
+  require("./userInfo")(app, db, upload, path, fs);
+  require("./userInfoWebSocket")(app, db, connection, pathname);
 });
 
 require("./routes/auth")(
